==> analysis/original/1) Predict Abstract Sample Size.ipynb <==
# Prototype Sample Size extraction using GPT

Goal: Test prompt engineering on the easiest use case: single 

import pandas as pd
from labelrepo.projects.participant_demographics import (
    get_participant_demographics,
)

subgroups = get_participant_demographics()

abstracts = pd.read_csv('data/abstracts.csv')

jerome_pd = subgroups[(subgroups.project_name == 'participant_demographics') & (subgroups.annotator_name == 'Jerome_Dockes')]

counts = jerome_pd.groupby('pmcid').count().reset_index()
single_group_pmcids = counts[counts['count'] == 1].pmcid

### Single group with Sample Size in Abstract

single_group = jerome_pd[jerome_pd.pmcid.isin(single_group_pmcids)]

single_group = pd.merge(single_group, abstracts[abstracts.pmcid.isin(single_group.pmcid)])

# Heuristic for finding in abstract studies
# Using fixed list below after manual checking
# single_group['in_abstract'] = \
#  single_group.apply(lambda x: (re.search(f"\D{x['count']}", x.abstract) != None), axis = 1)

# Manual list
not_in_abstract = [5548834, 4029023, 4318429, 5324609, 8752963, 5218407, 3147157, 2775905, 6344389, 8837589, 7430162, 7563756, 7156375, 4330553, 6989437, 6328158, 3409150, 3775427, 3483694, 6787094, 6528067, 3869649, 3183226, 6868994, 7002496, 6492297, 3780305]

single_group['in_abstract'] = single_group.pmcid.isin(not_in_abstract) == False

## Extract simple sample size

import openai
from templates import ZERO_SHOT_SAMPLE_SIZE_FUNCTION
from extract import extract_from_multiple

openai.api_key = open('/home/zorro/.keys/open_ai.key').read().strip()

predictions = extract_from_multiple(single_group.abstract.to_list(), 
                                    **ZERO_SHOT_SAMPLE_SIZE_FUNCTION, num_workers=2)

predictions['count'].astype('int', errors='ignore')

single_group['count'].astype('float')

# Normalize values
predictions['pmcid'] = single_group['pmcid']
single_group['count'] = single_group['count'].astype('float')

predictions['correct'] = single_group['count'] == predictions['count']

### Only look at those with sample size in abstract

in_abstract_predictions = predictions[single_group['in_abstract']]

# Percentage correct
in_abstract_predictions['correct'].mean()

# Incorrect predictions
in_abstract_predictions[in_abstract_predictions['correct'] == False]

Notes: First error is due to pulling out sample size for behavioral task, not fMRI task. 

Second error is due to study being a single subject study.

### Any hallucinations?

In all these cases where the information is not in the abstract, the model should have predicted either null, or 0

no_abstract_predictions = predictions[single_group['in_abstract'] == False]

# Predicted `n/a` or got correct answer (i.e. misclassified as not having info in abstract)
(pd.isna(no_abstract_predictions['count'])).mean()

# Studies where the model was wrong, and made a prediction
no_abstract_predictions[(pd.isna(no_abstract_predictions['count']) == False) & (no_abstract_predictions['correct'] == False)]

Overall notes: 

Errors:
- 2 error due to discrepancy between methods and abstract count (final n vs starting n)
- 1 n/a should be = 1
- 1 was for behavioral count, not fMRI count

False positives:
- Age was extracted instead of sample size (fixed after improving prompt)
- Returns 0 occasional instead of `n/a` (could be fixed in serialization). 


Excluding abstract/full-text discrepancies, error rate is 2/75
