==> analysis/original/4) Multi-group Prompts.ipynb <==
## Evaluating multi-group promps

import pandas as pd
import numpy as np
from labelrepo.projects.participant_demographics import get_participant_demographics
from labelrepo.database import get_database_connection

subgroups = get_participant_demographics(include_locations=True)
docs_info = pd.read_sql(
    "select pmcid, publication_year, title, text from document",
    get_database_connection(),

)

# Load multi group as well
jerome_pd = subgroups[(subgroups.project_name == 'participant_demographics') & \
                      (subgroups.annotator_name == 'Jerome_Dockes')]

counts = jerome_pd.groupby('pmcid').count().reset_index()
multi_group_pmcids = counts[counts['count'] > 1].pmcid
multi_group = jerome_pd[jerome_pd.pmcid.isin(multi_group_pmcids)]

# Get multi group docs
multi_group_docs = docs_info[docs_info.pmcid.isin(multi_group.pmcid)]

## Embed documents

import openai
openai.api_key = open('/home/zorro/.keys/open_ai.key').read().strip()

from embed import embed_pmc_articles, query_embeddings
texts = multi_group_docs[['pmcid', 'text']].to_dict(orient='records')
# embeddings = embed_pmc_articles(texts)
import pickle
# pickle.dump(embeddings, open('data/multi_group_embeddings.pkl', 'wb'))
embeddings = pickle.load(open('data/multi_group_embeddings.pkl', 'rb'))

embeddings = pd.DataFrame(embeddings)

## Extract group demographics across all docs 

from extract import extract_on_match
from templates import ZERO_SHOT_MULTI_GROUP
# sections, predictions = extract_on_match(embeddings, multi_group, **ZERO_SHOT_MULTI_GROUP, num_workers=3)
# pickle.dump(predictions, open('data/full_text_gpt4_multi_group.pkl', 'wb'))
predictions = pickle.load(open('data/full_text_gpt4_multi_group.pkl', 'rb'))

# Evaluation

# Subset annotation df to only include studies with body annotations
subset_cols = ['count', 'diagnosis', 'group_name', 'subgroup_name', 'male count',
       'female count', 'age mean', 'age minimum', 'age maximum',
       'age median', 'pmcid']
sub_multi_group = multi_group[multi_group.pmcid.isin(predictions.pmcid)][subset_cols].sort_values('pmcid')

from evaluate import evaluate_predictions, clean_predictions

clean_preds = clean_predictions(predictions)

ix_corr_n_groups, ix_more_groups, ix_less_groups = evaluate_predictions(predictions, sub_multi_group)

# With "cleaned up" predictions

_ = evaluate_predictions(clean_preds, sub_multi_group)

### Try on single group

single_group_embeddings = pickle.load(open('data/single_group_embeddings.pkl', 'rb'))
single_group_embeddings = pd.DataFrame(single_group_embeddings)

single_group_pmcids = counts[counts['count'] == 1].pmcid
single_group = jerome_pd[jerome_pd.pmcid.isin(single_group_pmcids)]

single_group_embeddings

single_sections, single_preds = extract_on_match(single_group_embeddings, single_group, **ZERO_SHOT_MULTI_GROUP, num_workers=3)
# single_preds = pickle.load(open('data/full_text_gpt4_single_group.pkl', 'rb'))

sub_single_group = single_group[single_group.pmcid.isin(single_preds.pmcid)][subset_cols].sort_values('pmcid')

single_preds

clean_single_preds = clean_predictions(single_preds)

ix_corr_n_groups_single, ix_more_groups_single, _= evaluate_predictions(single_preds, sub_single_group)

# With "cleaned up" predictions
_= evaluate_predictions(clean_single_preds, sub_single_group)

### Error types:

- Many times multiple rows get created for a single group study, each time a new "revised" sample size is mention (i.e. after exclusion)
- 2x the human annotation had less detai (i.e. total N instead sub group chars, that were combined into 1 group)

### Compare labels predicted vs given, on subset w/ matching group numbers

subset_pred_corrgroups[subset_pred_corrgroups.pmcid == 9435010]

subset_annot_corrgroups[subset_annot_corrgroups.pmcid == 9435010]

# Subset to only look at studies w/ same # of groups
subset_pred_corrgroups = predictions[predictions.pmcid.isin(ix_corr_n_groups)]
subset_annot_corrgroups = sub_multi_group[sub_multi_group.pmcid.isin(ix_corr_n_groups)]

annot_diag = subset_annot_corrgroups[(subset_annot_corrgroups.group_name == 'patients')][['count', 'pmcid', 'diagnosis']]
pred_diag = subset_pred_corrgroups[subset_pred_corrgroups.group_name == 'patients'][['count', 'pmcid', 'diagnosis']]
pd.merge(annot_diag, pred_diag, on=['pmcid', 'count'])

annot_diag = subset_annot_corrgroups[(subset_annot_corrgroups.group_name == 'healthy')][['count', 'pmcid', 'diagnosis']]
pred_diag = subset_pred_corrgroups[subset_pred_corrgroups.group_name == 'healthy'][['count', 'pmcid', 'diagnosis']]
pd.merge(annot_diag, pred_diag, on=['pmcid', 'count'])

### Other groups

subset_pred_corrgroups[subset_pred_corrgroups.group_name.isin(['healthy', 'patients']) == False][['count', 'pmcid', 'diagnosis', 'group_name']]


### Error types
- Will often fail to set name to `healthy` for controls.
   - Can probably come up with a whitelist of terms that can be translated to `healthy`:
      - ['control', 'volunteers', 'subjects', 'participants', 'adults'] + if diagnosis == `n/a` 

## Look at example of failures

### When model predicted more groups

sub_multi_group[sub_multi_group.pmcid.isin(ix_more_groups)].sort_values('pmcid')

predictions[predictions.pmcid.isin(ix_more_groups)].sort_values('pmcid')

sections[sections.pmcid == 9230060].iloc[0].content

### Error types
- Sometimes there is repeated information (for example, in Methods and Results). Not clear which section to "trust". In addition there are confusions about total N. 
- It seems to create a group if given the introduction, but it oftne lacks count. Which means we can probably drop those observations
- Final N and exclusion criteria always seems to confuse the models. In the case of multigroup prompt, instead of choosing one, it often will invent another group -- how to handle conflicting information?
- Also: it OFTEN, invents group_names other than `patients` or `healthy`. I think we can assume that if `diagnosis == NaN` or `healthy` then group_name can be corrected to `healthy`, and if there is a diagnosis then group_name should be patient. However, there will be rare times when something meaning healthy is given to diagnosis, and group name is something like 'control'.


Ideas for finding correction section:
- If can find "Methods" in body--- use that and search within that, or sections that follow.
- Also, should save subsection prior to splitting up-- save all headers to be able to select methods section
- If not, then search over entire document.
- Iterate until some information is generated, and iterate if there's no hits


### When model predicted fewer groups

sub_multi_group[sub_multi_group.pmcid.isin(ix_less_groups)].sort_values('pmcid')

predictions[predictions.pmcid.isin(ix_less_groups)].sort_values('pmcid')

sections[sections.pmcid == 9407088].iloc[0].content

### Error types
- Combining two groups into 1 (not necessarily wrong)-- happened 2x, usually combining control group w/ patient group
- Missed a dataset (abide)

Overall, not much we can do about this
