==> analysis/original/5) constrained fts + extract.ipynb <==
# Full Text Search + Extraction, pt ii

In this notebook I will reconnect FTS with the GPT extraction function, and evaluate the whole process across single and multi group studies.

To improve selection of the correct chunk, we'll use the following logic:

- Attempt to use heuristic to narrow down chunk, and search across entire document as fall back
- Feed one section to GPT, and if nothing is return, try next nearest section
- Evaluate using various output "cleaning rules"
- Try using GPT4 as well to see if there's improvements

import pandas as pd
import numpy as np
from labelrepo.projects.participant_demographics import get_participant_demographics

# Load annotations
subgroups = get_participant_demographics(include_locations=True)

# Load multi group as well
jerome_pd = subgroups[(subgroups.project_name == 'participant_demographics') & \
                      (subgroups.annotator_name == 'Jerome_Dockes')]

# Subset annotation df to only include studies with body annotations
subset_cols = ['count', 'diagnosis', 'group_name', 'subgroup_name', 'male count',
       'female count', 'age mean', 'age minimum', 'age maximum',
       'age median', 'pmcid']
jerome_pd_subset = jerome_pd[subset_cols].sort_values('pmcid')

jerome_pd_subset.pmcid.unique().shape

## Search and extract across all documents

import openai
import pickle

openai.api_key = open('/home/zorro/.keys/open_ai.key').read().strip()
all_embeddings = pickle.load(open('data/all_embeddings.pkl', 'rb'))

from extract import search_extract
from templates import ZERO_SHOT_MULTI_GROUP
from evaluate import evaluate_predictions, clean_predictions

query = 'How many participants or subjects were recruited for this study?' 

### No heuristic

# predictions_full_search = search_extract(all_embeddings, query, **ZERO_SHOT_MULTI_GROUP, num_workers=3)
# predictions_full_search.to_csv('data/predictions_full_search.csv', index=False)
predictions_full_search = pd.read_csv('data/predictions_full_search.csv')
# Clean predictions
clean_preds = clean_predictions(predictions_full_search)

predictions_full_search

import pprint
from publang.evaluate import compare_by_pmcid, score_columns

def evaluate_predictions(predictions, annotations):
    # Subset to only pmcids in predictions
    diff = set(set(annotations.pmcid.unique()) - set(predictions.pmcid.unique()))
    annotations = annotations[annotations.pmcid.isin(predictions.pmcid.unique())]

    pred_n_groups = predictions.groupby('pmcid').size()
    n_groups = annotations.groupby('pmcid').size()
    correct_n_groups = (n_groups == pred_n_groups)
    more_groups_pred = (n_groups < pred_n_groups)
    less_groups_pred = (n_groups > pred_n_groups)
    
    ix_corr_n_groups = correct_n_groups[correct_n_groups == True].index
    ix_more_groups =  more_groups_pred[more_groups_pred == True].index
    ix_less_groups = less_groups_pred[less_groups_pred == True].index

    print(f"Exact match # of groups: {correct_n_groups.mean():.2f}\n",
    f"More groups predicted: {more_groups_pred.mean():.2f}\n",
    f"Less groups predicted: {less_groups_pred.mean():.2f}\n",
    f"Missing pmcids: {diff}\n"
    )

    # Compare by columns (matched accuracy)
    print("Column wise comparison of predictions and annotations (error):\n")
    pprint.pprint(compare_by_pmcid(annotations, predictions))


    res_mean, res_sum, counts = score_columns(annotations, predictions, scoring='mpe')

    # Compare by columns (count of pmcids with overlap)
    print("\nPercentage response given by pmcid:\n")
    pprint.pprint(counts)
    

    # Compare by columns (summed mean percentage error on sum by pmcid)
    print("\nSummed Mean percentage error:\n")
    pprint.pprint(res_mean)

    # Compare by columns (summed mean percentage error on mean by pmcid)
    print("\nAveraged Mean percentage error:\n")
    pprint.pprint(res_sum)

    return ix_corr_n_groups, ix_more_groups, ix_less_groups

ix_corr_n_groups, ix_more_groups, ix_less_groups = evaluate_predictions(clean_preds, jerome_pd_subset)

### Heuristic - methods
uses heuristic to find methods section hen full text search using query

predictions_methods_fts = search_extract(all_embeddings, query, heuristic_strategy='methods', **ZERO_SHOT_MULTI_GROUP, num_workers=3)
predictions_methods_fts.to_csv('data/predictions_methods_fts.csv', index=False)
# predictions_methods_fts = pd.read_csv('data/predictions_methods_fts.csv')
# Clean predictions
predictions_methods_fts_clean = clean_predictions(predictions_methods_fts)

ix_corr_n_groups, ix_more_groups, ix_less_groups = evaluate_predictions(predictions_methods_fts_clean, jerome_pd_subset)

### Heurisic - demographics section

predictions_demographics_fts = search_extract(all_embeddings, query, heuristic_strategy='demographics', **ZERO_SHOT_MULTI_GROUP, num_workers=3)
predictions_demographics_fts.to_csv('data/predictions_demographics_fts.csv', index=False)
# predictions_demographics_fts = pd.read_csv('data/predictions_demographics_fts.csv')
# Clean predictions
predictions_demographics_fts_clean = clean_predictions(predictions_demographics_fts)

ix_corr_n_groups, ix_more_groups, ix_less_groups = evaluate_predictions(predictions_demographics_fts_clean, jerome_pd_subset)

# GPT4 - No Heuristic

# predictions_gpt4 = search_extract(all_embeddings, query, **ZERO_SHOT_MULTI_GROUP, num_workers=1, model_name='gpt-4')
predictions_gpt4 = pd.read_csv('data/predictions_gpt4.csv', index=False)

# Clean predictions
predictions_gpt4 = clean_predictions(predictions_gpt4)

ix_corr_n_groups, ix_more_groups, ix_less_groups = evaluate_predictions(predictions_gpt4, jerome_pd_subset)

pd.read_csv('data/pr

## Conclusion

Heuristic doesn't seem to improve things much, but does increase the chance that we don't find anything at all (when info is in Results section).

Could modify heuristic to fall back onto entire document if extraction comes  back null, but it doesn't seem to improve prediction much otherwise, so I'm not sure it's worthwhile. 

- GPT4 does marginally better but misses some docs entirely

### TODOs
- Refactor code into package
- 

# Manual results revision
