==> analysis/original/6) held out evaluation.ipynb <==
# Held out evaluation

This notebook will example result from held out evaluation

import pandas as pd

### Load Ground Truth
Papers annotated by Jerome

ground_truth = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsevaluation_labels.csv')
unique_ids = ground_truth.pmcid.unique()

### GPT predictions
- First load clean, and subset to include only those in ground truth (i.e. not those in training sample)
- Also load unclead data, to look at errors in papers that were cleaned up

gpt3_clean_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt_tokens-2000_clean.csv')
gpt3_all_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt_tokens-2000.csv')
embeddings = pd.read_parquet('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_embeddings_tokens-2000.parquet')

gpt4_clean_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt4_tokens-2000_clean.csv')
gpt4_all_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt4_tokens-2000.csv')

gpt4_turbo_clean_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt-4-1106-preview_tokens-2000_clean.csv')
gpt4_turbo_all_preds = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/eval_participant_demographics_gpt-4-1106-preview_tokens-2000.csv')

(gpt4_turbo_clean_preds[gpt4_turbo_clean_preds.pmcid == 8752963]['final'] == True).any()

def _keep_final(df):
    """ If within a PMCID any rows are annotated as final keep, otherwise return all"""
    if (df['final'] == True).any():
        return df[df['final'] == True]
    else:
        return df

def _merge_score(clean_preds, all_preds, ground_truth, unique_ids):
    clean_preds = clean_preds[clean_preds.pmcid.isin(unique_ids)]
    all_preds = all_preds[all_preds.pmcid.isin(unique_ids)]
    
    # For GPT-4 turbo look for final key
    if 'final' in clean_preds:
        clean_preds = clean_preds.groupby('pmcid').apply(_keep_final).reset_index(drop=True)
        all_preds = all_preds.groupby('pmcid').apply(_keep_final).reset_index(drop=True)
    
    clean_sum_count = clean_preds.groupby('pmcid').sum().reset_index()[['pmcid', 'count']]
    gt_sum_count = ground_truth.groupby('pmcid').sum().reset_index()[['pmcid', 'count']]
    merged = pd.merge(clean_sum_count, gt_sum_count, on='pmcid', )
    merged = merged.rename(columns={'count_y': 'true_count', 'count_x': 'prediction'})
    
    # Score prediction error
    merged['pe'] = abs((merged['true_count'] - merged['prediction']) / merged['true_count'])
    merged = merged.sort_values('pe')
    
    return clean_preds, all_preds, merged

gpt3_clean_preds, gpt3_all_preds, gpt3_merged = _merge_score(gpt3_clean_preds, gpt3_all_preds, ground_truth, unique_ids)

gpt4_clean_preds, gpt4_all_preds, gpt4_merged = _merge_score(gpt4_clean_preds, gpt4_all_preds, ground_truth, unique_ids)

gpt4_turbo_clean_preds, gpt4_turbo_all_preds, gpt4_turbo_merged = _merge_score(gpt4_turbo_clean_preds, gpt4_turbo_all_preds, ground_truth, unique_ids)

## Scores

gpt3_merged['pe'].median()

gpt3_merged['pe'].quantile(0.75)

gpt3_merged['pe'].mean()

gpt4_merged['pe'].median()

gpt4_merged['pe'].quantile(0.75)

gpt4_merged['pe'].mean()

Excluding one outlier, GPT-4 perform slightly better:

gpt3_merged.iloc[0:-1]['pe'].mean()

gpt4_merged.iloc[0:-1]['pe'].mean()

gpt4_turbo_merged['pe'].median()

gpt4_turbo_merged['pe'].mean()

gpt4_turbo_merged[-10:]

## Explore

tid = 3672681

gpt4_turbo_all_preds[gpt4_turbo_all_preds.pmcid == tid]

gpt4_turbo_all_preds[gpt4_turbo_all_preds.pmcid == tid]

ground_truth[ground_truth.pmcid == tid]

content = embeddings[(embeddings.pmcid == tid) & (embeddings.start_char == gpt4_all_preds[gpt4_all_preds.pmcid == tid].iloc[0].start_char)].iloc[0].content

content

### Observations

GPT-3 & 4 are both very good at extracting sample size, with GPT-4 being a bit better sometimes (althuogh more amibitious and extracting more groups).

Challenge is if given more than 1 group, how to select fMRI group?
Often counts are full counts, and not including exclusions

GPT-4 often reports both inital and final counts
