==> analysis/original/7)  participant demographics (gpt) over large datsaet.ipynb <==
import pandas as pd
import numpy as np
import seaborn as sns

gpt3_all = pd.read_csv('../fmri_participant_demographics/data/outputs/demographicsgpt/all_participant_demographics_gpt_tokens-4000_clean.csv')

gpt3_all

gpt3_all.diagnosis.value_counts()

#### How to do named entity recognition on this list?

We need a set of possible values, ontology (disease ontology)
Ask GPT to avoid acronyms? Go back a second time and ask to match to set?

gpt3_all.diagnosis.value_counts().index.tolist()

gpt3_all.group_name.value_counts()
